Chroma core/chrome 536 (#568)
## Description of changes

Please let me know if anything needs to be changed or done differently!
We can certainly continue to make efforts to make our JS client more
structured and maintainable.

*Summary:*
 - Improvements & Bug fixes
	 - Broke `client/js/src/index.ts` down and put them in separate files
- Created `IEmbeddedFunction` interface to improve the maintainability
and extensibility of the client code, making it easier to add new
embedding function types in the future
	 - Replaced `CallableFunction` with `IEmbeddingFunction`
- Closes #536
 - New functionality
	 - None

## Test plan
Refactored tests 
**Note:** the Collection tests have been failing on main when I run
using `yarn test`, did we know about this?

## Documentation Changes
Not sure, couldn't find anything?
