+++ clients/js/src/embeddings/IEmbeddingFunction.ts
@@ -1,0 +1,3 @@
+export interface IEmbeddingFunction {
+    generate(texts: string[]): Promise<number[][]>;
+}
